requirements.txt
streamlit
geopandas
pandas
shapely.geometry
numpy
folium
